requirements.txt
asgiref==3.8.1
Django==5.0.6
django-allauth==0.63.0
pillow==10.3.0
sqlparse==0.5.0
typing_extensions==4.11.0
